Animation fixes + more porting
Fixing gg2c's GGInfo fixed the animation offsets
continuing to stub out code to get things to compile

+++ project/source/engine/scroll_utils.c
@@ -6,7 +6,7 @@
 
 u16 ScrollUtils_mapColumnToBuild;
 
-void ScrollUtils_buildColumn(u8* buffer)
+void ScrollUtils_buildColumn(u16* buffer)
 {
 /*
