+++ project/source/engine/scroll_utils.h
@@ -1,11 +1,12 @@
 #ifndef SCROLL_UTILS_INCLUDE_H
 #define SCROLL_UTILS_INCLUDE_H
 
+#include "genesis.h"
 #include "engine/base_types.h"
 #include "engine/map_types.h"
 #include "engine/object_types.h"
 
-void ScrollUtils_buildColumn(u8* buffer);
+void ScrollUtils_buildColumn(u16* buffer);
 void ScrollUtils_InitTilemap(void);
 void ScrollUtils_InitTilemapStrip(void);
